requirements.txt
streamlit==1.31.0
face_recognition==1.3.0
pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0